+++ apps/docs/server/assets/snapshots/express/public/script.js
@@ -1,5 +1,9 @@
 function doSomething() {
+  console.log('Fetching message...')
   fetch('/api/message')
     .then((response) => response.text())
-    .then((data) => (document.querySelector('.response').innerHTML = data))
+    .then((data) => {
+      console.log('Received message')
+      document.querySelector('.response').innerHTML = data
+    })
 }