+++ packages/sandbox/src/types/index.ts
@@ -1,3 +1,4 @@
+export * from './Console.ts'
 export * from './EditorViews.ts'
 export * from './Panel.ts'
 export * from './Sandbox.ts'